Introduce ISqLiteJsonCacheManager with factory for cache manager and other uses
