+++ src/Speckle.Sdk/Serialisation/V2/SerializeProcessFactory.cs
@@ -2,6 +2,7 @@ using Speckle.Sdk.Helpers;
 using Speckle.Sdk.Logging;
 using Speckle.Sdk.Serialisation.V2.Receive;
 using Speckle.Sdk.Serialisation.V2.Send;
+using Speckle.Sdk.SQLite;
 using Speckle.Sdk.Transports;
 
 namespace Speckle.Sdk.Serialisation.V2;
@@ -27,7 +28,8 @@ public class SerializeProcessFactory(
   ISdkActivityFactory activityFactory,
   IBaseChildFinder baseChildFinder,
   IObjectSerializerFactory objectSerializerFactory,
-  IObjectDeserializerFactory objectDeserializerFactory
+  IObjectDeserializerFactory objectDeserializerFactory,
+  ISqLiteJsonCacheManagerFactory sqLiteJsonCacheManagerFactory
 ) : ISerializeProcessFactory
 {
   public ISerializeProcess CreateSerializeProcess(
@@ -37,11 +39,11 @@ public class SerializeProcessFactory(
     IProgress<ProgressArgs>? progress
   )
   {
-    var sqliteSendCacheManager = new SQLiteSendCacheManager(streamId);
+    var sqLiteJsonCacheManager = sqLiteJsonCacheManagerFactory.CreateFromStream(streamId);
     var serverObjectManager = new ServerObjectManager(speckleHttp, activityFactory, url, streamId, authorizationToken);
     return new SerializeProcess(
       progress,
-      sqliteSendCacheManager,
+      sqLiteJsonCacheManager,
       serverObjectManager,
       baseChildFinder,
       objectSerializerFactory
@@ -55,10 +57,10 @@ public class SerializeProcessFactory(
     IProgress<ProgressArgs>? progress
   )
   {
-    var sqliteSendCacheManager = new SQLiteReceiveCacheManager(streamId);
+    var sqLiteJsonCacheManager = sqLiteJsonCacheManagerFactory.CreateFromStream(streamId);
     var serverObjectManager = new ServerObjectManager(speckleHttp, activityFactory, url, streamId, authorizationToken);
 
-    var objectLoader = new ObjectLoader(sqliteSendCacheManager, serverObjectManager, progress);
+    var objectLoader = new ObjectLoader(sqLiteJsonCacheManager, serverObjectManager, progress);
     return new DeserializeProcess(progress, objectLoader, objectDeserializerFactory);
   }
 }
