+++ tests/Speckle.Sdk.Serialization.Testing/Program.cs
@@ -11,6 +11,7 @@ using Speckle.Sdk.Serialisation.V2;
 using Speckle.Sdk.Serialisation.V2.Receive;
 using Speckle.Sdk.Serialisation.V2.Send;
 using Speckle.Sdk.Serialization.Testing;
+using Speckle.Sdk.SQLite;
 
 const bool skipCacheReceive = false;
 const bool skipCacheSendCheck = true;
@@ -46,7 +47,8 @@ var factory = new SerializeProcessFactory(
   serviceProvider.GetRequiredService<ISdkActivityFactory>(),
   new BaseChildFinder(new BasePropertyGatherer()),
   new ObjectSerializerFactory(new BasePropertyGatherer()),
-  new ObjectDeserializerFactory()
+  new ObjectDeserializerFactory(),
+  serviceProvider.GetRequiredService<ISqLiteJsonCacheManagerFactory>()
 );
 var process = factory.CreateDeserializeProcess(new Uri(url), streamId, token, progress);
 var @base = await process.Deserialize(rootId, default, new(skipCacheReceive)).ConfigureAwait(false);
