+++ tests/Speckle.Sdk.Serialization.Tests/DummySqLiteReceiveManager.cs
@@ -1,15 +1,14 @@
-﻿using Speckle.Sdk.Dependencies.Serialization;
-using Speckle.Sdk.Serialisation.V2;
+﻿using Speckle.Sdk.SQLite;
 
 namespace Speckle.Sdk.Serialization.Tests;
 
-public class DummySqLiteReceiveManager(Dictionary<string, string> savedObjects) : ISQLiteReceiveCacheManager
+public class DummySqLiteReceiveManager(Dictionary<string, string> savedObjects) : ISqLiteJsonCacheManager
 {
   public string? GetObject(string id) => savedObjects.GetValueOrDefault(id);
 
-  public void SaveObject(BaseItem item) => throw new NotImplementedException();
+  public void SaveObject(string id, string json) => throw new NotImplementedException();
 
-  public void SaveObjects(List<BaseItem> item) => throw new NotImplementedException();
+  public void SaveObjects(IEnumerable<(string id, string json)> items) => throw new NotImplementedException();
 
   public bool HasObject(string objectId) => throw new NotImplementedException();
 }
