+++ tests/Speckle.Sdk.Serialization.Tests/DummySqLiteSendManager.cs
@@ -1,13 +1,13 @@
-﻿using Speckle.Sdk.Dependencies.Serialization;
-using Speckle.Sdk.Serialisation.V2;
+﻿using Speckle.Sdk.SQLite;
 
 namespace Speckle.Sdk.Serialization.Tests;
 
-public class DummySqLiteSendManager : ISQLiteSendCacheManager
+public class DummySqLiteSendManager : ISqLiteJsonCacheManager
 {
   public string? GetObject(string id) => throw new NotImplementedException();
+  public void SaveObject(string id, string json) => throw new NotImplementedException();
+
+  public void SaveObjects(IEnumerable<(string id, string json)> items) => throw new NotImplementedException();
 
   public bool HasObject(string objectId) => throw new NotImplementedException();
-
-  public void SaveObjects(List<BaseItem> items) => throw new NotImplementedException();
 }
